master - NEW: adding timeframe buttons

+++ app/controllers/frames_controller.rb
@@ -10,7 +10,7 @@ class FramesController < ApplicationController
 
   def get_cell_reports
     respond_to do |format|
-      format.json { render json: Frame.find(params[:id]).json_reports(params[:timeframe].to_i)}
+      format.json { render json: Frame.find(params[:id]).json_reports(params[:timeframe].to_f)}
     end
   end
